@@ -25,11 +25,14 @@
   },
   "scripts": {
     "start": "cross-env ENV=dev electron .",
-    "watch": "webpack-dev-server --config webpack.config.renderer.dev.js",
-    "watch-main": "webpack --config webpack.config.main.prod.js --watch --mode development",
+    "watch": "concurrently \"npm run watch-renderer\" \"npm run watch-main\" \"npm run watch-preloads\"",
+    "watch-renderer": "node server.js",
+    "watch-main": "webpack --config webpack.config.main.js --watch --mode development",
+    "watch-preloads": "webpack --config webpack.config.preloads.js --watch --mode development",
     "build-renderer": "webpack --config webpack.config.renderer.prod.js",
-    "build-main": "webpack --config webpack.config.main.prod.js",
-    "build": "npm run build-main && npm run build-renderer",
+    "build-main": "webpack --config webpack.config.main.js --mode production",
+    "build-preloads": "webpack --config webpack.config.preloads.js --mode production",
+    "build": "npm run build-main && npm run build-preloads && npm run build-renderer",
     "compile-win32": "npm run build && electron-builder -w -p always",
     "compile-darwin": "npm run build && electron-builder -m -p always",
     "compile-linux": "npm run build && electron-builder -l -p always",
@@ -42,18 +45,18 @@
   "devDependencies": {
     "@types/chrome": "0.0.71",
     "@types/gsap": "1.19.2",
-    "@types/node": "10.5.4",
+    "@types/node": "10.5.6",
     "@types/react": "16.4.7",
     "@types/react-dom": "16.0.6",
     "awesome-typescript-loader": "5.2.0",
-    "babel-core": "6.26.3",
-    "babel-loader": "7.1.5",
+    "concurrently": "3.6.1",
     "cross-env": "5.2.0",
     "dexie": "2.0.4",
-    "electron": "2.0.5",
-    "electron-builder": "20.26.0",
+    "electron": "2.0.6",
+    "electron-builder": "20.26.1",
     "electron-react-devtools": "0.5.3",
-    "eslint": "5.2.0",
+    "electron-reload": "Sential/electron-reload",
+    "eslint": "5.3.0",
     "eslint-config-airbnb": "17.0.0",
     "eslint-plugin-import": "2.13.0",
     "eslint-plugin-jsx-a11y": "6.1.1",
@@ -66,20 +69,20 @@
     "prettier": "1.14.0",
     "prettier-eslint": "8.8.2",
     "prettier-eslint-cli": "4.7.1",
-    "react": "16.4.1",
-    "react-dom": "16.4.1",
+    "react": "16.4.2",
+    "react-dom": "16.4.2",
     "react-hot-loader": "4.3.4",
-    "styled-components": "3.3.3",
+    "styled-components": "3.4.0",
     "typescript": "3.0.1",
     "typescript-eslint-parser": "17.0.1",
     "url-loader": "1.0.1",
-    "webpack": "4.16.3",
+    "webpack": "4.16.4",
     "webpack-cli": "3.1.0",
     "webpack-dev-server": "3.1.5",
-    "webpack-merge": "4.1.4"
+    "webpack-merge": "4.1.4",
+    "wexond-package-manager": "1.4.5"
   },
   "dependencies": {
-    "electron-updater": "3.0.3",
-    "wexond-package-manager": "1.4.5"
+    "electron-updater": "3.0.3"
   }
 }
